# 10x faster `new Buffer`