--- a/src/bun.js/bindings/JSDOMConvertBufferSource+JSBuffer.h
+++ b/src/bun.js/bindings/JSDOMConvertBufferSource+JSBuffer.h
@@ -8,24 +8,4 @@
 
 namespace WebCore {
 
-struct IDLJSBuffer : IDLTypedArray<WebCore::JSBuffer> {
-};
-
-template<> struct JSConverter<IDLJSBuffer> {
-    static constexpr bool needsState = true;
-    static constexpr bool needsGlobalObject = true;
-
-    template<typename U>
-    static JSC::JSValue convert(JSC::JSGlobalObject& lexicalGlobalObject, JSDOMGlobalObject& globalObject, const U& value)
-    {
-        return toJS(&lexicalGlobalObject, &globalObject, Detail::getPtrOrRef(value));
-    }
-
-    template<typename U>
-    static JSC::JSValue convertNewlyCreated(JSC::JSGlobalObject& lexicalGlobalObject, JSDOMGlobalObject& globalObject, U&& value)
-    {
-        return convert(lexicalGlobalObject, globalObject, std::forward<U>(value));
-    }
-};
-
 }