… src/bun.js/bindings/JSBuffer.cpp
… src/bun.js/builtins/cpp/JSBufferPrototypeBuiltins.cpp
… src/bun.js/builtins/js/JSBufferPrototype.js
… test/bun.js/node-stream-uint8array.test.ts