```zig
❯ ~/.bun/bin/bun buffer-create.mjs 10 # Before
cpu: Apple M1 Max
runtime: bun 0.4.0 (arm64-darwin)

benchmark                       time (avg)             (min … max)       p75       p99      p995
------------------------------------------------------------------ -----------------------------
new Buffer(0)               236.38 ns/iter (191.03 ns … 717.36 ns) 203.87 ns 621.13 ns  657.7 ns
new Buffer(10)              234.89 ns/iter (193.65 ns … 676.88 ns) 200.53 ns 531.55 ns 660.81 ns
Buffer.alloc(10)            239.54 ns/iter  (194.36 ns … 676.1 ns) 206.95 ns  656.1 ns 665.66 ns
Buffer.allocUnsafe(10)      234.59 ns/iter (192.95 ns … 627.66 ns) 203.14 ns 518.39 ns 542.68 ns
Buffer.allocUnsafe(24_000)  662.21 ns/iter  (471.5 ns … 861.91 ns) 736.33 ns 861.91 ns 861.91 ns
Buffer.alloc(24_000)          1.28 µs/iter     (1.13 µs … 1.62 µs)   1.32 µs   1.62 µs   1.62 µs


❯ bun buffer-create.mjs 10 # After
cpu: Apple M1 Max
runtime: bun 0.4.1 (arm64-darwin)

benchmark                       time (avg)             (min … max)       p75       p99      p995
------------------------------------------------------------------ -----------------------------
new Buffer(0)                17.65 ns/iter  (15.81 ns … 283.28 ns)  16.55 ns   62.7 ns  76.86 ns
new Buffer(10)                22.5 ns/iter  (17.64 ns … 400.41 ns)  22.01 ns  65.08 ns  76.51 ns
Buffer.alloc(10)             20.07 ns/iter   (18.57 ns … 256.3 ns)  18.92 ns  62.16 ns  80.38 ns
Buffer.allocUnsafe(10)       21.33 ns/iter   (19.82 ns … 243.4 ns)  20.39 ns  64.04 ns  75.88 ns
Buffer.allocUnsafe(24_000)  413.74 ns/iter (245.16 ns … 646.53 ns) 458.61 ns  614.7 ns 646.53 ns
Buffer.alloc(24_000)        985.37 ns/iter   (830.01 ns … 1.24 µs)   1.03 µs   1.24 µs   1.24 µs


❯ node buffer-create.mjs 10 # Node.js for comparison
cpu: Apple M1 Max
runtime: node v19.2.0 (arm64-darwin)

benchmark                       time (avg)             (min … max)       p75       p99      p995
------------------------------------------------------------------ -----------------------------
new Buffer(0)                32.76 ns/iter  (23.98 ns … 101.06 ns)  32.65 ns  71.75 ns  83.17 ns
new Buffer(10)               40.86 ns/iter  (33.31 ns … 258.96 ns)  39.83 ns 120.11 ns 139.52 ns
Buffer.alloc(10)             32.44 ns/iter   (31.93 ns … 39.31 ns)  32.44 ns  36.53 ns  36.92 ns
Buffer.allocUnsafe(10)       27.92 ns/iter    (25.4 ns … 35.76 ns)  30.31 ns  32.21 ns  33.93 ns
Buffer.allocUnsafe(24_000)   436.2 ns/iter  (265.5 ns … 978.32 ns) 498.92 ns 846.75 ns 978.32 ns
Buffer.alloc(24_000)          1.03 µs/iter    (920.1 ns … 1.36 µs)   1.05 µs   1.36 µs   1.36 µs
```